requirements.txt
fastapi
uvicorn[standard]
streamlit
requests
pydantic
python-dotenv
sqlalchemy
google-generativeai
